renamed periodic report and removed the actor version
