--- gcm4/src/main/java/plugins/reports/support/PeriodicReport2.java
@@ -1,234 +1,0 @@
-package plugins.reports.support;
-
-import java.util.function.BiConsumer;
-
-import nucleus.Event;
-import nucleus.NucleusError;
-import nucleus.ReportContext;
-import util.errors.ContractException;
-
-/**
- * The abstract base class for reports that aggregate reporting aligned to a
- * {@link ReportPeriod}. The periodic report continually schedules reporting on
- * a regular cycle, invoking the flush method and allowing descendant
- * implementors to release collected data. This report registers via the context
- * to be alerted when the simulation terminates and will perform a final flush
- * invocation. This can result in the duplication of time values for last data
- * released.
- *
- *
- */
-public abstract class PeriodicReport2  {
-
-	private ReportContext reportContext;
-
-	/**
-	 * Creates the periodic report from the given report period
-	 * 
-	 * @throws ContractException
-	 *             <li>if the report period is null</li>
-	 */
-	public PeriodicReport2(ReportId reportId, ReportPeriod reportPeriod) {
-		if (reportPeriod == null) {
-			throw new ContractException(ReportError.NULL_REPORT_PERIOD);
-		}
-		this.reportPeriod = reportPeriod;
-
-		if (reportId == null) {
-			throw new ContractException(ReportError.NULL_REPORT_ID);
-		}
-		this.reportId = reportId;
-	}
-
-	/*
-	 * Assume a daily report period and let it be overridden
-	 */
-	private ReportPeriod reportPeriod = ReportPeriod.DAILY;
-
-	private ReportId reportId;
-
-	/*
-	 * The day value to be used in report lines
-	 */
-	private Integer reportingDay = 0;
-
-	/*
-	 * The hour value to be used in report lines
-	 */
-	private Integer reportingHour = 0;
-
-	/**
-	 * Adds the time field column(s) to the given {@link ReportHeaderBuilder} as
-	 * appropriate to the {@link ReportPeriod} specified during construction.
-	 * 
-	 * DAILY : Day
-	 * 
-	 * HOURLY : Day, Hour
-	 * 
-	 * END_OF_SIMULATION has no header additions
-	 */
-	protected final ReportHeader.Builder addTimeFieldHeaders(ReportHeader.Builder reportHeaderBuilder) {
-		switch (reportPeriod) {
-		case DAILY:
-			reportHeaderBuilder.add("day");
-			break;
-		case END_OF_SIMULATION:
-			// do nothing
-			break;
-		case HOURLY:
-			reportHeaderBuilder.add("day");
-			reportHeaderBuilder.add("hour");
-			break;
-		default:
-			throw new RuntimeException("unknown report period " + reportPeriod);
-		}
-		return reportHeaderBuilder;
-	}
-
-	protected final ReportId getReportId() {
-		return reportId;
-	}
-
-	/**
-	 * Places the current reporting day and report hour on the report as
-	 * appropriate to the {@link ReportPeriod} specified during construction.
-	 *
-	 */
-	protected final void fillTimeFields(final ReportItem.Builder reportItemBuilder) {
-
-		switch (reportPeriod) {
-		case DAILY:
-			reportItemBuilder.addValue(reportingDay);
-			break;
-		case END_OF_SIMULATION:
-			// do nothing
-			break;
-		case HOURLY:
-			reportItemBuilder.addValue(reportingDay);
-			reportItemBuilder.addValue(reportingHour % 24);
-			break;
-		default:
-			throw new RuntimeException("unknown report period " + reportPeriod);
-		}
-	}
-
-	/**
-	 * Subscribes to simulation close. Initializes periodic flushing of report
-	 * contents with the first flush scheduled for one time period from
-	 * simulation start. Descendant implementors of PeriodicReport must invoke
-	 * super.init().
-	 * 
-	 * @throws ContractException
-	 *             <li>if the report context is null</li>
-	 * 
-	 */
-	public void init(ReportContext reportContext) {
-
-		if (reportContext == null) {
-			throw new ContractException(ReportError.NULL_CONTEXT);
-		}
-		this.reportContext = reportContext;
-
-		reportContext.subscribeToSimulationClose(this::close);
-
-		if (reportPeriod != ReportPeriod.END_OF_SIMULATION) {
-			setNextPlanTime();
-			reportContext.addPlan(this::executePlan, nextPlanTime);
-		}
-	}
-
-	private void close(final ReportContext reportContext) {
-		if (lastFlushTime == null || reportContext.getTime() > lastFlushTime) {
-			lastFlushTime = reportContext.getTime();
-			flush(reportContext);
-		}
-	}
-
-	/**
-	 * Provides descendant implementors the opportunity to releases report items
-	 * from the data stored during the time since the last invocation of
-	 * flush().
-	 */
-	protected abstract void flush(final ReportContext reportContext);
-
-	private double nextPlanTime;
-	private Double lastFlushTime;
-
-	private void setNextPlanTime() {
-		switch (reportPeriod) {
-		case DAILY:
-			nextPlanTime = (reportingDay + 1);
-			break;
-		case HOURLY:
-			nextPlanTime = reportingDay + (double) (reportingHour + 1) / 24;
-			break;
-		default:
-			throw new RuntimeException("unhandled report period " + reportPeriod);
-		}
-	}
-
-	/**
-	 * Returns a wrapped version of the given consumer that will ensure proper
-	 * flushing when events are received at the same time as a flushing plan,
-	 * but happen to execute before the plan. Descendant implementors of
-	 * PeriodicReport should use this wrapper when subscribing to events.
-	 */
-	private <T extends Event> BiConsumer<ReportContext, T> getFlushingConsumer(BiConsumer<ReportContext, T> eventConsumer) {
-		return (c, t) -> {
-			if (c.getTime() >= nextPlanTime) {
-				if (this.reportPeriod != ReportPeriod.END_OF_SIMULATION) {
-					if (lastFlushTime == null || c.getTime() > lastFlushTime) {
-						lastFlushTime = c.getTime();
-						flush(c);
-					}
-				}
-			}
-			eventConsumer.accept(c, t);
-		};
-	}
-
-	private void executePlan(final ReportContext reportContext) {
-		
-		if (lastFlushTime == null || reportContext.getTime() > lastFlushTime) {
-			lastFlushTime = reportContext.getTime();
-			flush(reportContext);
-		}
-
-		switch (reportPeriod) {
-		case DAILY:
-			reportingDay++;
-			break;
-		case HOURLY:
-			reportingHour++;
-			if (reportingHour == 24) {
-				reportingHour = 0;
-				reportingDay++;
-			}
-			break;
-		default:
-			throw new RuntimeException("unhandled report period " + reportPeriod);
-		}
-
-		setNextPlanTime();
-		reportContext.addPlan(this::executePlan, nextPlanTime);
-
-	}
-
-	/**
-	 * Subscribes the report to the given event filter via the actor context
-	 * while enforcing the flushing of report items as needed. Events of the
-	 * type T are processed by the event filter. If the event passes the filter
-	 * the event will be consumed by the supplied event consumer.
-	 * 
-	 * 
-	 * @throws ContractException
-	 *             <li>{@link NucleusError#NULL_EVENT_FILTER} if the event
-	 *             filter is null
-	 *             <li>{@link NucleusError#NULL_EVENT_CONSUMER} if the event
-	 *             consumer is null
-	 */
-	protected final <T extends Event> void subscribe(Class<T> eventClass, BiConsumer<ReportContext, T> eventConsumer) {
-		reportContext.subscribe(eventClass, getFlushingConsumer(eventConsumer));
-	}
-
-}
